# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/vehicles": {
      "post": {
        "operationId": "VehiclesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVehicleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Vehicle created.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 201
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add a new vehicle",
        "tags": [
          "Vehicles"
        ]
      },
      "get": {
        "operationId": "VehiclesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of vehicles."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user's vehicles",
        "tags": [
          "Vehicles"
        ]
      }
    },
    "/vehicles/{id}": {
      "get": {
        "operationId": "VehiclesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Vehicle ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vehicle details."
          },
          "404": {
            "description": "Vehicle not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get vehicle details",
        "tags": [
          "Vehicles"
        ]
      },
      "patch": {
        "operationId": "VehiclesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Vehicle ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVehicleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Vehicle updated."
          },
          "404": {
            "description": "Vehicle not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a vehicle",
        "tags": [
          "Vehicles"
        ]
      },
      "delete": {
        "operationId": "VehiclesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Vehicle ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vehicle deleted."
          },
          "404": {
            "description": "Vehicle not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a vehicle",
        "tags": [
          "Vehicles"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Documentation",
    "description": "Swagger pour ton backend NestJS",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "BadRequestResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "message": {
            "type": "object",
            "example": "Bad Request"
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
      "UnauthorizedResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 401
          },
          "message": {
            "type": "object",
            "example": "Unauthorized"
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
      "NotFoundResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 404
          },
          "message": {
            "type": "object",
            "example": "Not Found"
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
...
      "CreateVehicleDto": {
        "type": "object",
        "properties": {
          "brand": {
            "type": "string",
            "example": "Toyota"
          },
          "model": {
            "type": "string",
            "example": "Corolla"
          },
          "plateNumber": {
            "type": "string",
            "example": "ABC-1234"
          },
          "seats": {
            "type": "number",
            "example": 4,
            "default": 4
          },
          "imageUrl": {
            "type": "string",
            "example": "https://example.com/car.jpg"
          }
        },
        "required": [
          "brand",
          "model",
          "seats"
        ]
      },
      "UpdateVehicleDto": {
        "type": "object",
        "properties": {
          "brand": {
            "type": "string",
            "example": "Toyota"
          },
          "model": {
            "type": "string",
            "example": "Corolla"
          },
          "plateNumber": {
            "type": "string",
            "example": "ABC-1234"
          },
          "seats": {
            "type": "number",
            "example": 4,
            "default": 4
          },
          "imageUrl": {
            "type": "string",
            "example": "https://example.com/car.jpg"
          }
        }
      },
...
    }
  }
}